requirements.txt
arabic-reshaper==2.1.4
et-xmlfile==1.1.0
fpdf==1.7.2
future==0.18.2
numpy==1.23.3
openpyxl==3.0.10
pandas==1.5.0
Pillow==9.2.0
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2022.2.1
reportlab==3.6.11
six==1.16.0
xlrd==2.0.1
